Split course and server user references into two files.
